% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_95CI.R
\name{OR_95CI}
\alias{OR_95CI}
\title{Odds Ratio Confidence Interval Function}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{numeric vector}

\item{se}{numeric vector}

\item{siglevel}{significance level}

\item{roundto}{significant digits}
}
\value{
CI
}
\description{
A function that returns the confidence interval of the estimated coefficients from logistic regression
}
\examples{
data(toydata)
log.reg <- glm( y~x1+x2, data = toydata, family = "binomial")
coefs<- summary(log.reg)$coefficients[2,1]
se <- summary(log.reg)$coefficients[2,2]
OR_95CI(coef=coefs, se=se, siglevel=0.05, roundto = 2)
}
\author{
Szewei Wang
}
