% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Toydata}
\format{
A data frame with 100 rows and 3 variables:
\describe{
\item{y}{y variable}
\item{x1}{covariate}
\item{x2}{covariate}
}
}
\usage{
toydata
}
\description{
Test Data for logistic regression example
}
\keyword{datasets}
